requirements.txt
asgiref==3.3.1
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.3.0
mysqlclient==2.0.1
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
